requirements.txt
# Dependências do projeto
python-dotenv>=0.21.0
requests>=2.28.0
pandas>=1.5.0
numpy>=2.2.0
setuptools-80.7.1
# tqdm>=4.65.0  # opcional, para barras de progresso
